backup

+++ src/main/java/com/mycompany/qlthanhvien/DAL/DAL_ThanhVien.java
@@ -4,7 +4,6 @@
  */
 package com.mycompany.qlthanhvien.DAL;
 
-import com.mycompany.qlthanhvien.BLL.BLL_ThanhVien;
 import com.mycompany.qlthanhvien.BLL.ThanhVien;
 import java.util.ArrayList;
 import java.util.List;
@@ -12,57 +11,95 @@ import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 
-
-
 /**
  *
  * @author Admin
  */
 public class DAL_ThanhVien {
-    
+
     private Session session;
-    
-    public DAL_ThanhVien (){
-         session = HibernateUtils.getSessionFactory().openSession();
+
+    public DAL_ThanhVien() {
+        session = HibernateUtils.getSessionFactory().openSession();
     }
-   
-     public List<ThanhVien> getThanhvien() {
-         
-         List<ThanhVien> thanhvien;
-         session.beginTransaction();
+
+    public List<ThanhVien> getThanhvien() {
+
+        List<ThanhVien> thanhvien;
+        session.beginTransaction();
         thanhvien = this.session.createQuery("FROM ThanhVien").list();
         session.getTransaction().commit();
         return thanhvien;
     }
-    public List<ThanhVien> SearchThanhvien(String column, String value ) {
-         
-         List<ThanhVien> thanhvien;
-         session.beginTransaction();
-    
-        thanhvien = this.session.createQuery("FROM ThanhVien WHERE " + column + " = "+ value).list();
+
+    public List<ThanhVien> SearchThanhvien(String column, String value) {
+
+        List<ThanhVien> thanhvien;
+        session.beginTransaction();
+
+        thanhvien = this.session.createQuery("FROM ThanhVien WHERE " + column + " = " + value).list();
         session.getTransaction().commit();
         return thanhvien;
     }
-     
-    public void addThanhVien(ThanhVien thanhvien)
-    {
+
+    public void addThanhVien(ThanhVien thanhvien) {
         session.beginTransaction();
-        session.save(thanhvien);   
+        session.save(thanhvien);
         session.getTransaction().commit();
-        
+
     }
-    public void updateThanhVien(ThanhVien thanhvien)
-    {
-       session.beginTransaction();
-       session.update(thanhvien);
-       session.getTransaction().commit();
-    }
-    public void deleteThanhVien(ThanhVien thanhvien)
-    {
+
+    public void updateThanhVien(ThanhVien thanhvien) {
         session.beginTransaction();
+        session.update(thanhvien);
+        session.getTransaction().commit();
+    }
+
+    public void deleteThanhVien(ThanhVien thanhvien) throws Exception {
+        session.beginTransaction();
+        // Kiểm tra xem mã thành viên có tồn tại trong bảng xuly hay không
+        Long countXuLy = (Long) session.createQuery("SELECT COUNT(*) FROM XuLy WHERE MaTV = :maTV")
+                .setParameter("maTV", thanhvien.getMaTV())
+                .uniqueResult();
+        if (countXuLy > 0) {
+            throw new Exception("Không thể xóa thành viên do tồn tại trong bảng xuly.");
+        }
+        // Nếu không có ràng buộc khóa ngoại, thực hiện xóa
         session.delete(thanhvien);
         session.getTransaction().commit();
     }
+
+    public boolean deleteThanhVienByYear(int year) {
+        session.beginTransaction();
+        List<ThanhVien> thanhViensToDelete = session.createQuery("FROM ThanhVien WHERE SUBSTRING(MaTV, 3, 2) = :year")
+                .setParameter("year", String.valueOf(year))
+                .list();
+
+        boolean canDelete = true;
+        for (ThanhVien thanhVien : thanhViensToDelete) {
+            // Kiểm tra xem MaTV của ThanhVien có tồn tại trong bảng ThongTinsd hay không
+            Long countThongTinsd = (Long) session.createQuery("SELECT COUNT(*) FROM ThongTinSD WHERE MaTV = :maTV")
+                    .setParameter("maTV", thanhVien.getMaTV())
+                    .uniqueResult();
+            // Kiểm tra xem MaTV của ThanhVien có tồn tại trong bảng XuLy hay không
+            Long countXuLy = (Long) session.createQuery("SELECT COUNT(*) FROM XuLy WHERE MaTV = :maTV")
+                    .setParameter("maTV", thanhVien.getMaTV())
+                    .uniqueResult();
+
+            // Nếu MaTV tồn tại trong cả hai bảng ThongTinsd và XuLy, không thể xóa
+            if (countThongTinsd > 0 || countXuLy > 0) {
+                canDelete = false;
+            } else {
+                // Nếu không tồn tại trong cả hai bảng, thực hiện xóa thành viên
+                session.delete(thanhVien);
+            }
+        }
+
+        session.getTransaction().commit();
+
+        return canDelete; // Trả về true nếu có thể xóa, ngược lại trả về false
+    }
+
     public boolean mergeThanhVien(ThanhVien tv) {
         Transaction transaction = null;
         try {
@@ -78,40 +115,42 @@ public class DAL_ThanhVien {
             return false;
         }
     }
+
     public static void main(String[] args) {
         DAL_ThanhVien dAL_ThanhVien = new DAL_ThanhVien();
-        
-       //  Lấy ds
-        List<ThanhVien> tvs = dAL_ThanhVien.getThanhvien();
-        for (int i = 0;i< tvs.size(); i++) { 
-            System.out.println(tvs.get(i).getMaTV());
-            
-        }
-        
 
-//        add
-//        ThanhVien thanhVien = new ThanhVien();
-//        thanhVien.setMaTV(1243);
-//        thanhVien.setHoten("abc");
-//        thanhVien.setEmail("emailcom");
-//        thanhVien.setPassword("123123h");
-//        dAL_ThanhVien.addThanhVien(thanhVien);
-        
-// update
-//        ThanhVien thanhVien = new ThanhVien();
-//        thanhVien.setMaTV(1243);
-//        thanhVien.setHoten("loc hoang");
-//        thanhVien.setEmail("emailcom");
-//        thanhVien.setPassword("123123h");
-//        dAL_ThanhVien.updateThanhVien(thanhVien);
-        
-//        delete 
-//        ThanhVien thanhVien = new ThanhVien();
-//        thanhVien.setMaTV(1243);
-//         dAL_ThanhVien.deleteThanhVien(thanhVien);
-        
-          
+        // Lấy ds
+        List<ThanhVien> tvs = dAL_ThanhVien.SearchThanhvien("MaTV", "1124420005");
+        for (int i = 0; i < tvs.size(); i++) {
+            System.out.println(tvs.get(i).getMaTV());
+
+        }
     }
-    
-    
+
+    public List<Integer> getDistinctYears() {
+        List<Integer> years = new ArrayList<>();
+        session.beginTransaction();
+        List<ThanhVien> thanhvien = session.createQuery("FROM ThanhVien").list();
+        session.getTransaction().commit();
+        for (ThanhVien tv : thanhvien) {
+            String maTV = String.valueOf(tv.getMaTV());
+            String yearString = maTV.substring(2, 4);
+            int year = Integer.parseInt(yearString);
+            if (!years.contains(year)) {
+                years.add(year);
+            }
+        }
+        return years;
+    }
+
+    public List<ThanhVien> getThanhVienByYear(int year) {
+        List<ThanhVien> thanhViens;
+        session.beginTransaction();
+        thanhViens = session.createQuery("FROM ThanhVien WHERE SUBSTRING(MaTV, 3, 2) = :year")
+                .setParameter("year", String.valueOf(year))
+                .list();
+        session.getTransaction().commit();
+        return thanhViens;
+    }
+
 }
